backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
PyMuPDF>=1.23.0
python-multipart>=0.0.6
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
python-dotenv>=1.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0 

# Caching
redis>=4.5.0 

# System Monitoring
psutil>=5.9.0 

# File Processing
PyMuPDF>=1.23.0 